change: relocate ocr_test
