+++ hrp/pkg/uixt/service_vedem_test.go
@@ -86,3 +86,14 @@ func TestTapUIWithScreenshot(t *testing.T) {
 		t.Fatal(err)
 	}
 }
+
+func TestDriverExtOCR(t *testing.T) {
+	driverExt, err := iosDevice.NewDriver(nil)
+	checkErr(t, err)
+
+	point, err := driverExt.FindScreenText("抖音")
+	checkErr(t, err)
+
+	t.Logf("point.X: %v, point.Y: %v", point.X, point.Y)
+	driverExt.Driver.TapFloat(point.X, point.Y-20)
+}
